arc-sentinel-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
supabase==2.3.0
httpx==0.24.1
